backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
alembic==1.12.1

# ML/NLP
sentence-transformers==2.7.0
tiktoken==0.5.2
numpy==1.26.2

# Document Processing
pypdf==3.17.1
beautifulsoup4==4.12.2
markdownify==0.11.6

# Search
rank-bm25==0.2.2

# LLM
openai==1.3.7
anthropic==0.7.8
ollama==0.1.7

# Task Queue
celery[redis]==5.3.4
redis==4.6.0  # Fixed: Changed from 5.0.1 to 4.6.0 for compatibility

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
black==23.11.0